move cluster metrics to module
